added  movie-card.jsx and movie-list.jsx
